+++ test/pleroma/upload/filter/analyze_metadata_test.exs
@@ -2,9 +2,9 @@
 # Copyright © 2017-2021 Pleroma Authors <https://pleroma.social/>
 # SPDX-License-Identifier: AGPL-3.0-only
 
-defmodule Pleroma.Upload.Filter.SetMetaTest do
+defmodule Pleroma.Upload.Filter.AnalyzeMetadataTest do
   use Pleroma.DataCase, async: true
-  alias Pleroma.Upload.Filter.SetMeta
+  alias Pleroma.Upload.Filter.AnalyzeMetadata
 
   test "adds the image dimensions" do
     upload = %Pleroma.Upload{
@@ -14,6 +14,6 @@ defmodule Pleroma.Upload.Filter.SetMetaTest do
       tempfile: Path.absname("test/fixtures/image.jpg")
     }
 
-    assert {:ok, :filtered, %{width: 1024, height: 768}} = SetMeta.filter(upload)
+    assert {:ok, :filtered, %{width: 1024, height: 768}} = AnalyzeMetadata.filter(upload)
   end
 end